Make caster header-only
