+++ cmake/nanobind-config.cmake
@@ -128,7 +128,6 @@ function (nanobind_build_library TARGET_NAME)
     ${NB_DIR}/src/nb_type.cpp
     ${NB_DIR}/src/nb_enum.cpp
     ${NB_DIR}/src/nb_ndarray.cpp
-    ${NB_DIR}/src/nb_datetime.cpp
     ${NB_DIR}/src/common.cpp
     ${NB_DIR}/src/error.cpp
     ${NB_DIR}/src/trampoline.cpp
